artisan command created to send reminders
